AMT-21 Enable assignment grading screen on Altitude / M31
As a result of a change in core Moodle in M31, the assignment grading screen
would not display properly. The fix overrides the relative positioning set
on the page body by the slidebars style sheet. Override is only for the
assignment grader screen.

AMT-21: Fixed broken layout for assignment grader screen.